ADD: createManyProducts

+++ api/services/product.service.js
@@ -1,14 +1,20 @@
+const xlsx = require('xlsx');
+
 const ProductRepository = require('../db/repositories/product.repository');
+const SupplierRepository = require('../db/repositories/supplier.repository');
 const CloudinaryService = require('./cloudinary.service');
 
 const { validateFile } = require('../utils/multer');
 const emptyUploadsDirectory = require('../utils/emptyUploadsDirectory');
 
 const CLOUDINARY_PRODUCTS_FOLDER = 'products';
+const NO_IMAGE_URL = 'https://res.cloudinary.com/dw6ri09rd/image/upload/v1722170457/products/no_image_qxtmba.jpg';
+const NO_IMAGE_ID = 'no_image_qxtmba';
 
 class ProductService {
   constructor() {
     this._repository = new ProductRepository()
+    this._supplierRepository = new SupplierRepository()
     this._cloudinaryService = new CloudinaryService()
   }
 
@@ -49,6 +55,10 @@ class ProductService {
     ) {
       throw new Error('Please complete all fields')
     }
+
+    if (typeof cost != 'number' || typeof price != 'number') {
+      throw new Error('Cost and Price must be in number format')
+    }
   }
 
   async createProduct(product) {
@@ -75,8 +85,85 @@ class ProductService {
     }
   }
 
-  // TODO
-  async createManyProducts(products) { }
+  async parseWorkbookToProduct(products) {
+    const parsedProducts = [];
+
+    for (const product of products) {
+      const {
+        nombre,
+        categoria,
+        marca,
+        articulo,
+        modelo,
+        sku,
+        ean,
+        descripcion,
+        costo,
+        precio,
+        proveedor,
+        seccion
+      } = product;
+
+      const supplier = await this._supplierRepository.findSupplierByName(proveedor);
+      if (!supplier) {
+        throw new Error(`No supplier with name ${proveedor}`);
+      }
+
+      const parsedProduct = {
+        name: nombre,
+        category: categoria,
+        brand: marca,
+        article: articulo,
+        model: modelo,
+        sku,
+        eanCode: ean,
+        description: descripcion,
+        cost: costo,
+        price: precio,
+        supplier: supplier._id,
+        section: seccion,
+        image: NO_IMAGE_URL,
+        imageId: NO_IMAGE_ID
+      };
+      parsedProducts.push(parsedProduct);
+    }
+
+    return parsedProducts;
+  }
+
+  async validateWorkbook(file) {
+    const { Sheets: sheets } = file
+    const sheetsNames = Object.keys(sheets);
+    if (sheetsNames.length != 1) {
+      throw new Error('File must have exactly one sheet');
+    }
+
+
+    const sheet = sheetsNames[0];
+    const sheetProducts = xlsx.utils.sheet_to_json(sheets[sheet]);
+    const products = await this.parseWorkbookToProduct(sheetProducts);
+
+    products.forEach((prod) => {
+      this.validateProduct(prod);
+    });
+
+    return products;
+  }
+
+  async createManyProducts(file) {
+    const { path } = file;
+    const workbook = xlsx.readFile(path)
+
+    try {
+      const products = await this.validateWorkbook(workbook);
+      await this._repository.createMany(products);
+      emptyUploadsDirectory();
+    } catch (error) {
+      emptyUploadsDirectory();
+      throw error;
+    }
+
+  }
 
   async updateProduct(productId, product) {
     try {
@@ -99,7 +186,9 @@ class ProductService {
       const oldProduct = await this._repository.updateById(productId, updateData);
 
       const { imageId } = oldProduct;
-      await this._cloudinaryService.deleteImage(imageId);
+      if (imageId != NO_IMAGE_ID) {
+        await this._cloudinaryService.deleteImage(imageId);
+      }
 
       emptyUploadsDirectory();
     } catch (error) {
@@ -120,7 +209,9 @@ class ProductService {
         throw new Error(`Product with id ${productId} does not exist`);
       }
 
-      await this._cloudinaryService.deleteImage(deletedProduct.imageId);
+      if (deletedProduct.imageId != imageId) {
+        await this._cloudinaryService.deleteImage(deletedProduct.imageId);
+      }
     } catch (error) {
       throw error;
     }
